refactor(sync-service): move sync services to correct files
