+++ src/lib/utils/constants.ts
@@ -1,0 +1,1 @@
+export const MIN_SYNC_INTERVAL = 5 * 60 * 1000; // 5 minutes
