+++ src/routes/+layout.server.ts
@@ -5,14 +5,10 @@ import { checkRegistryHealth } from '$lib/utils/api/health';
 import { initDatabase, syncFromRegistry, incrementalSync } from '$lib/services/database';
 import { db } from '$lib/services/database/connection';
 import { runMigrations } from '$lib/services/database/migrations.ts';
+import { getLastSyncTime, updateLastSyncTime } from '$lib/services/database';
+import { MIN_SYNC_INTERVAL } from '$lib/utils/constants';
 
 const logger = Logger.getInstance('LayoutServer');
-
-// Min time between syncs (in milliseconds)
-const MIN_SYNC_INTERVAL = 5 * 60 * 1000; // 5 minutes
-
-// Store last sync time in memory (will reset on server restart)
-let lastSyncTime = 0;
 
 export async function load({ url }) {
 	// Proceed with regular functionality for non-test environments
@@ -42,7 +38,6 @@ export async function load({ url }) {
 			await incrementalSync(registryData.repositories, { forceFullSync });
 
 			// Update last sync time in both memory and DB
-			lastSyncTime = currentTime;
 			updateLastSyncTime(currentTime);
 
 			logger.info('Registry data synced to database');
@@ -85,26 +80,6 @@ function ensureSettingsTable() {
     `
 		).run();
 		logger.info('Settings table created');
-	}
-}
-
-// Get the last sync time from the database
-function getLastSyncTime() {
-	try {
-		return db.prepare('SELECT value FROM settings WHERE key = ?').get('last_sync_time');
-	} catch (error) {
-		logger.error('Error getting last sync time:', error);
-		return null;
-	}
-}
-
-// Update the last sync time in the database
-function updateLastSyncTime(timestamp: number) {
-	try {
-		const stmt = db.prepare('INSERT OR REPLACE INTO settings (key, value) VALUES (?, ?)');
-		stmt.run('last_sync_time', timestamp);
-	} catch (error) {
-		logger.error('Error updating last sync time:', error);
 	}
 }
